requirements.txt
torch>=2.1.0
transformers>=4.41.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
cryptography>=42.0.0
python-dotenv>=1.0.1
konlpy>=0.6.0; platform_system != "Windows"  # optional; may require Java
PyYAML>=6.0.1
